requirements.txt
# requirements.txt
# Core Dependencies
aiohttp==3.9.1
aiosqlite==0.19.0
asyncio-throttle==1.0.2
flask==2.3.3
pyyaml==6.0.1
schedule==1.2.0

# Data Processing
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2

# API and Web
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Utilities
python-dateutil==2.8.2
pytz==2023.3
ujson==5.8.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
flake8==6.1.0

# Production
gunicorn==21.2.0
supervisor==4.2.5